arena/setup/requirements.txt
cdp-sdk
python-dotenv
